02-3 타입 변환

+++ chap02/src/sec02/exam02/ByteExample.java
@@ -7,7 +7,7 @@ public class ByteExample {
 		byte var3 = 0;
 		byte var4 = 30;
 		byte var5 = 127;
-		byte var6 = 128;	//컴파일 에러(Type mismatch:cannot convert from int to byte)
+		int/*byte*/ var6 = 128;	//컴파일 에러(Type mismatch:cannot convert from int to byte)
 		
 		System.out.println(var1);
 		System.out.println(var2);
